@@ -8,8 +8,8 @@ import android.content.Context
 import android.content.Intent
 import android.graphics.Color
 import android.os.Build
-import android.support.v4.app.NotificationCompat
 import android.util.Log
+import androidx.core.app.NotificationCompat
 import org.jetbrains.anko.doAsync
 import vladyslavpohrebniakov.notgood.MySqlHelper
 import vladyslavpohrebniakov.notgood.R
